backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
matplotlib
seaborn
python-multipart
python-dotenv
google-generativeai
Pillow
